+++ docs/manager/graphql-reference/v2-schema.graphql
@@ -198,6 +198,11 @@ input StringFilter {
   endsWith: String = null
   equals: String = null
   notEquals: String = null
+  iContains: String = null
+  iStartsWith: String = null
+  iEndsWith: String = null
+  iEquals: String = null
+  iNotEquals: String = null
 }
 
 type Subscription {
