+++ src/ai/backend/manager/api/gql/base.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from typing import Optional
+
 import strawberry
 
 
@@ -10,3 +12,18 @@ class ByteSize:
     """
 
     pass
+
+
+@strawberry.input
+class StringFilter:
+    contains: Optional[str] = None
+    starts_with: Optional[str] = None
+    ends_with: Optional[str] = None
+    equals: Optional[str] = None
+    not_equals: Optional[str] = None
+
+    i_contains: Optional[str] = strawberry.field(name="iContains", default=None)
+    i_starts_with: Optional[str] = strawberry.field(name="iStartsWith", default=None)
+    i_ends_with: Optional[str] = strawberry.field(name="iEndsWith", default=None)
+    i_equals: Optional[str] = strawberry.field(name="iEquals", default=None)
+    i_not_equals: Optional[str] = strawberry.field(name="iNotEquals", default=None)
